pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "langchain>=0.3.26",
    "langgraph>=0.5.3",
    "pillow>=11.3.0",
    "telegrinder>=0.5.1",
    "uuid>=1.30",
    # Authentication and OAuth
    "django-allauth>=0.60.1",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.1",
    # Google APIs
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
    # Social integrations
    "requests>=2.31.0",
    "python-telegram-bot>=20.7",
    # Utilities
    "python-decouple>=3.8",
    "django-environ>=0.11.2",
    "cryptography>=41.0.0",
]
